+++ debug-info.js
@@ -1,7 +1,10 @@
 /* global process, __resourceQuery */
-if( process && process.env.NODE_ENV !== 'production' ) {
+if( /(^\?|\&)lazy(\&|$)/.test( __resourceQuery ) ) {
    module.exports = require( '!!bundle-loader?lazy&name=debug-info!./lib/index?debug&entries&' + __resourceQuery.substr( 1 ) + '!./dummy.js' );
 }
-else {
+else if( process && process.env && process.env.NODE_ENV === 'production' ) {
    module.exports = { info: 'debug info not available in production' };
+}
+else {
+   module.exports = require( './lib/index?debug&entries&' + __resourceQuery.substr( 1 ) + '!./dummy.js' );
 }