+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Last Changes
 
+
+## v2.0.0-rc.0
+
 - [#28](https://github.com/LaxarJS/laxar-loader/issues/28): updated README.md
 - [#19](https://github.com/LaxarJS/laxar-loader/issues/19): singular and plural options (`?flow=` vs. `?flows[]=`) may now be used interchangeably
 - [#18](https://github.com/LaxarJS/laxar-loader/issues/18): themes and layouts must now have descriptor `.json` files